requirements.txt
certifi==2023.7.22
chardet==5.1.0
docutils==0.20.1
filestack-python==3.5.0
idna==3.4
Kivy==2.2.1
Kivy-Garden==0.1.5
Pygments==2.15.1
requests==2.31.0
trafaret==2.1.1
urllib3==2.0.4
python-dotenv==1.0.0
